Updated SQL examples
