+++ examples/python/quickstart_sql.py
@@ -19,57 +19,56 @@ spark = SparkSession.builder \
     .getOrCreate()
 
 # Clear any previous runs
+spark.sql("DROP TABLE IF EXISTS " + tableName)
+spark.sql("DROP TABLE IF EXISTS newData")
+
 try:
-    spark.sql("DROP TABLE IF EXISTS " + tableName)
-except:
-    pass
+    # Create a table
+    print("############# Creating a table ###############")
+    spark.sql("CREATE TABLE %s(id LONG) USING delta" % tableName)
+    spark.sql("INSERT INTO %s VALUES 0, 1, 2, 3, 4" % tableName)
 
-# Create a table
-print("############# Creating a table ###############")
-spark.range(0, 5).write.format("delta").mode("overwrite").saveAsTable(tableName)
+    # Read the table
+    print("############ Reading the table ###############")
+    spark.sql("SELECT * FROM %s" % tableName).show()
 
-# Read the table
-print("############ Reading the table ###############")
-spark.sql("SELECT * FROM %s" % tableName).show()
+    # Upsert (merge) new data
+    print("########### Upsert new data #############")
+    spark.sql("CREATE TABLE newData(id LONG) USING parquet")
+    spark.sql("INSERT INTO newData VALUES 3, 4, 5, 6")
 
-# Upsert (merge) new data
-print("########### Upsert new data #############")
-newData = spark.range(0, 20).write.format("delta").mode("overwrite").saveAsTable("newData")
+    spark.sql('''MERGE INTO {0} USING newData
+            ON {0}.id = newData.id
+            WHEN MATCHED THEN 
+              UPDATE SET {0}.id = newData.id
+            WHEN NOT MATCHED THEN INSERT *
+        '''.format(tableName))
 
-spark.sql('''MERGE INTO {0} USING newData
-        ON {0}.id = newData.id
-        WHEN MATCHED THEN 
-          UPDATE SET {0}.id = newData.id
-        WHEN NOT MATCHED THEN INSERT *
-    '''.format(tableName))
+    spark.sql("SELECT * FROM %s" % tableName).show()
 
-spark.sql("SELECT * FROM %s" % tableName).show()
-
-# Update table data
-print("########## Overwrite the table ###########")
-data = spark.range(5, 10)
-data.write.format("delta").mode("overwrite").saveAsTable(tableName)
-spark.sql("SELECT * FROM %s" % tableName).show()
+    # Update table data
+    print("########## Overwrite the table ###########")
+    spark.sql("INSERT OVERWRITE %s select * FROM (VALUES 5, 6, 7, 8, 9) x (id)" % tableName)
+    spark.sql("SELECT * FROM %s" % tableName).show()
 
 
-# Update every even value by adding 100 to it
-print("########### Update to the table(add 100 to every even value) ##############")
-spark.sql("UPDATE {0} SET id = (id + 100) WHERE (id % 2 == 0)".format(tableName))
-spark.sql("SELECT * FROM %s" % tableName).show()
+    # Update every even value by adding 100 to it
+    print("########### Update to the table(add 100 to every even value) ##############")
+    spark.sql("UPDATE {0} SET id = (id + 100) WHERE (id % 2 == 0)".format(tableName))
+    spark.sql("SELECT * FROM %s" % tableName).show()
 
-# Delete every even value
-print("######### Delete every even value ##############")
-spark.sql("DELETE FROM {0} WHERE (id % 2 == 0)".format(tableName))
-spark.sql("SELECT * FROM %s" % tableName).show()
+    # Delete every even value
+    print("######### Delete every even value ##############")
+    spark.sql("DELETE FROM {0} WHERE (id % 2 == 0)".format(tableName))
+    spark.sql("SELECT * FROM %s" % tableName).show()
 
+    # Read old version of data using time travel
+    print("######## Read old data using time travel ############")
+    df = spark.read.format("delta").option("versionAsOf", 0).table(tableName)
+    df.show()
 
-# Read old version of data using time travel
-print("######## Read old data using time travel ############")
-df = spark.read.format("delta").option("versionAsOf", 0).table(tableName)
-df.show()
-
-# cleanup
-try:
+finally:
+    # cleanup
     spark.sql("DROP TABLE " + tableName)
-except:
-    pass
+    spark.sql("DROP TABLE IF EXISTS newData")
+    spark.stop()
